refactor: move buildPermitCallData to utils
